Removed unused methods and tests
