+++ tests/test_cache.py
@@ -1,30 +1,16 @@
 from unittest import mock
 
+import pytest
+from gql.transport.exceptions import TransportServerError
+
+from contract_tests.test_proposal_system_behaviour_mocked import (
+    VALID_INSTRUMENT_LIST,
+    generate_standard_mock,
+)
 from example_data import get_ymir_example_data
 from yuos_query.cache import Cache
+from yuos_query.exceptions import ConnectionException
 from yuos_query.proposal_system import ProposalSystem
-
-VALID_INSTRUMENT_LIST = [
-    {"id": 4, "shortCode": "YMIR", "description": "Our test beamline", "name": "YMIR"},
-    {
-        "id": 3,
-        "shortCode": "asztalos",
-        "description": "test\n",
-        "name": "Asztalos Instrument",
-    },
-    {
-        "id": 2,
-        "shortCode": "s adasd",
-        "description": "d asdas",
-        "name": "Test instrument 2",
-    },
-    {
-        "id": 1,
-        "shortCode": "sd ad",
-        "description": "d asdasd",
-        "name": "Test instrument 1",
-    },
-]
 
 YMIR_EXAMPLE_DATA = get_ymir_example_data()
 
@@ -73,3 +59,29 @@ def test_cache_gives_dictionary_of_proposals():
     assert cache.cached_proposals["471120"].samples[1].number == 1
     assert cache.cached_proposals["471120"].samples[1].density == (0, "g/cm*3")
     assert cache.cached_proposals["471120"].samples[1].mass_or_volume == (0, "µg")
+
+
+SOME_URL = "https://something.com"
+SOME_TOKEN = "not_a_real_token"
+VALID_PROPOSAL_ID = "169"
+YMIR_INFO = (4, "YMIR")
+
+
+def test_issue_with_getting_instrument_data_from_system_raises_correct_exception_type():
+    mocked_impl = generate_standard_mock()
+    mocked_impl.get_instrument_data.side_effect = TransportServerError("oops")
+
+    cache = Cache(SOME_URL, SOME_TOKEN, "YMIR", implementation=mocked_impl)
+    with pytest.raises(ConnectionException):
+        cache.refresh()
+
+
+def test_issue_with_getting_proposal_data_from_system_raises_correct_exception_type():
+    mocked_impl = generate_standard_mock()
+    mocked_impl.get_instrument_data.return_value = VALID_INSTRUMENT_LIST
+    mocked_impl.get_proposals_including_samples_for_instrument.side_effect = (
+        TransportServerError("oops")
+    )
+    cache = Cache(SOME_URL, SOME_TOKEN, "YMIR", implementation=mocked_impl)
+    with pytest.raises(ConnectionException):
+        cache.refresh()
